Games are now linked to the Owner useruser

+++ controllers/GameController.js
@@ -14,8 +14,17 @@ const getGames = async(req, res) => {
 }
 
 const addGame = async (req, res) => {
+    const userId = req.userId
     try{
-        const game = await Games.create(req.body)
+        const {title, genre, year} = req.body
+
+        const game = await Games.create({
+            title: title,
+            genre: genre,
+            year: year,
+            ownerId: userId
+        })
+
         res.status(200).json(game)
     }
     catch (error){
@@ -27,8 +36,14 @@ const addGame = async (req, res) => {
 const editGame = async (req, res) => {
     try{
         const{id} = req.params;
-        console.log(id)
+        const owner = await Games.findById(id)
+        if(owner.ownerId != req.userId)
+        {
+            return res.status(404).json({message: "Usuário não está ligado a esse jogo."})
+        }
         const game = await Games.findByIdAndUpdate(id, req.body);
+        console.log(game.ownerId)
+
         if(!game) {
             return res.status(404).json({message: "Jogo não encontrado"})
         }
@@ -42,7 +57,12 @@ const editGame = async (req, res) => {
 
 const deleteGame = async (req, res) => {
     try{
-        const {id} = await req.body;
+        const {id} = await req.params;
+        const owner = await Games.findById(id)
+        if(owner.ownerId != req.userId)
+        {
+            return res.status(404).json({message: "Usuário não está ligado a esse jogo."})
+        }
         const game = await Games.findByIdAndDelete(id);
         if(!game){
             res.status(404)
